chore: move theme.ts to themes dir
